Changes made after testing
DriveTrain likely wasn't working because it was set to speed AND the Enable and Disable methods weren't called in Robot.cpp
Arm moved without stopping when it hit the limit switch. Method changed to give it joystick values. Limit switch issue will hopefully be fixed with this, otherwise it's an electrical problem.
Pneumatic systems open and close set to be on different buttons because having them on the same button didn't work in practice. Buttons chosen may need to be changed, will discuss at meeting.